Исправлена проблема с событиями и добавлены новые типы ссылок. Правда теперь враппер не работает с библиотекой ниже 0.3.0

+++ ScEngineNet/Native/WScMemoryContext.cs
@@ -11,7 +11,7 @@ namespace ScEngineNet.Native
         /// <summary>
         /// The identifier
         /// </summary>
-        internal ushort Id;
+        internal uint Id;
         /// <summary>
         /// The access levels
         /// </summary>
